+++ VLPT/modeling_transfo_xl_utilities_16.py
@@ -54,7 +54,7 @@ class ProjectedAdaptiveLogSoftmax(nn.Module):
         if div_val == 1:
             for i in range(len(self.cutoffs)):
                 if d_proj != d_embed:
-                    self.out_projs.append(nn.Parameter(torch.FloatTensor(d_proj, d_embed)))
+                    self.out_projs.append(nn.Parameter(torch.BFloat16Tensor(d_proj, d_embed)))
                 else:
                     self.out_projs.append(None)
 
@@ -64,7 +64,7 @@ class ProjectedAdaptiveLogSoftmax(nn.Module):
                 l_idx, r_idx = self.cutoff_ends[i], self.cutoff_ends[i + 1]
                 d_emb_i = d_embed // (div_val**i)
 
-                self.out_projs.append(nn.Parameter(torch.FloatTensor(d_proj, d_emb_i)))
+                self.out_projs.append(nn.Parameter(torch.BFloat16Tensor(d_proj, d_emb_i)))
 
                 self.out_layers.append(nn.Linear(d_emb_i, r_idx - l_idx))
